+++ src/pages/Main/components/CardList/CardText.js
@@ -1,8 +1,7 @@
 import transferImg from 'assets/train_chien_woman.png';
-import roadFindImg from 'assets/map_open.png';
 import busTimeImg from 'assets/smartphone_map_app_man.png';
 
-function SlideCard({ status }) {
+function CardText({ type }) {
 	function Card1Text() {
 		return (
 			<div className='main-page__slide-info'>
@@ -16,27 +15,6 @@ function SlideCard({ status }) {
 						<p>번거롭지 않으세요?</p>
 					</div>
 					<p>열차와 버스 시간표를 한 번에 조회해보세요.</p>
-				</div>
-			</div>
-		);
-	}
-
-	function Card2Text() {
-		return (
-			<div className='main-page__slide-info'>
-				<div className='main-page__slide-info-img-container'>
-					<img src={roadFindImg} alt='밀양 길찾기 이미지' />
-				</div>
-				<div className='main-page__slide-info-text'>
-					<div className='main-page__slide-point-texts'>
-						<p className='point'>밀양시내,</p>
-						<p>알려드릴게요</p>
-					</div>
-					<p>
-						목적지에서
-						<br />
-						가장 가까운 정류장을 알려드려요.
-					</p>
 				</div>
 			</div>
 		);
@@ -67,11 +45,10 @@ function SlideCard({ status }) {
 
 	return (
 		<>
-			{status === 0 ? <Card1Text /> : ''}
-			{status === 1 ? <Card2Text /> : ''}
-			{status === 2 ? <Card3Text /> : ''}
+			{type === 0 ? <Card1Text /> : ''}
+			{type === 1 ? <Card3Text /> : ''}
 		</>
 	);
 }
 
-export default SlideCard;
+export default CardText;
