+++ src/pages/Main/components/CardList/main.js
@@ -1,11 +1,11 @@
 import { Link } from 'react-router-dom';
-import SlideCard from 'components/MainPage/SlideCard';
+import CardText from './CardText';
 import Button from 'components/Common/Button';
 
-function CardList({ status, link, content }) {
+function CardItem({ type, link, content }) {
 	return (
 		<li className='main-page__slide-item'>
-			<SlideCard status={status} />
+			<CardText type={type} />
 			<div className='main-page__slide-item-button-container'>
 				<Link to={link}>
 					<Button size={'middle'}>{content}</Button>
@@ -15,4 +15,4 @@ function CardList({ status, link, content }) {
 	);
 }
 
-export default CardList;
+export default CardItem;
