+++ src/pages/TrainTransferResult/main.js
